+++ backend/src/main/java/sd/faturas/FaturasRequestDTO.java
@@ -2,7 +2,7 @@ package sd.faturas;
 
 import sd.pedidos.Pedidos;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -11,7 +11,7 @@ public record FaturasRequestDTO
                 Pedidos pedidoId,
                 String numeroFatura,
                 LocalDate dataEmissao,
-                BigInteger valorTotal,
+                BigDecimal valorTotal,
                 LocalDate dataVencimento,
                 String status,
                 LocalDateTime createdAt
