+++ backend/src/main/java/sd/pedidoItens/PedidoItensRequestDTO.java
@@ -3,17 +3,17 @@ package sd.pedidoItens;
 import core.produtos.Produtos;
 import sd.pedidos.Pedidos;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record PedidoItensRequestDTO
         (
                 Pedidos pedidoId,
                 Produtos produtoId,
-                BigInteger quantidade,
-                BigInteger valorUnitario,
-                BigInteger valorTotal,
-                BigInteger desconto,
+                BigDecimal quantidade,
+                BigDecimal valorUnitario,
+                BigDecimal valorTotal,
+                BigDecimal desconto,
                 LocalDateTime createdAt
         ) {
 }
