+++ backend/src/main/java/sd/pedidoItens/PedidoItensResponseDTO.java
@@ -2,7 +2,7 @@ package sd.pedidoItens;
 
 import core.produtos.Produtos;
 import sd.pedidos.Pedidos;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record PedidoItensResponseDTO
@@ -10,10 +10,10 @@ public record PedidoItensResponseDTO
                 Integer id,
                 Pedidos pedidoId,
                 Produtos produtoId,
-                BigInteger quantidade,
-                BigInteger valorUnitario,
-                BigInteger valorTotal,
-                BigInteger desconto,
+                BigDecimal quantidade,
+                BigDecimal valorUnitario,
+                BigDecimal valorTotal,
+                BigDecimal desconto,
                 LocalDateTime createdAt
         )
     {
